Ugly hack to make GC restart the Teensy
I've found a work-around to get GC talking to the Teensy. A macro discussed [here](https://forum.pjrc.com/threads/24304-_reboot_Teensyduino()-vs-_restart_Teensyduino()?p=45253&viewfull=1#post45253) works to restart the Teensy without resetting the USB port. I needed a way for GC to trigger that when making a connection to a Teensy. A quick-and-dirty approach is to define a '$'  command (I picked '$@') to send just after making the serial connection which the Teensy would catch.